Added initial version of the "Quotes on Design" Alexa Skill
This is my submission for the Amazon June Giveaway of an Echo Dot. This Alexa Skill is similar to the Android app I created, Quotr. By requesting a new quote or asking to be inspired, Alexa will respond with a quote and its corresponding author. Again, I used the "Quotes on Design" (http://quotesondesign.com/) API.

I do plan on improving this Skill to allow using "slots", basically parameters used when making a request, such as "ask Quotes on Design for a quote by {author}".